requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
langchain==0.3.14
langchain-groq==0.2.3
pydantic==2.10.4
python-dotenv==1.0.1
